Replace routing with npm page
